Migrate rate interval from panel to target. Update README
