+++ README.md
@@ -17,9 +17,10 @@ The linter implements the following rules:
 * `template-job-rule` - Checks that the dashboard has a templated job and instance.
 * `template-label-promql-rule` - Checks that the dashboard templated labels have proper PromQL expressions.
 * `panel-datasource-rule` - Checks that each panel uses the templated datasource.
-* `panel-promql-rule` - Checks that each panel uses a valid PromQL query.
-* `panel-rate-interval-rule` - Checks that each panel uses $__rate_interval.
-* `panel-job-instance-rule` - Checks that every PromQL query has job and instance matchers.
+* `target-promql-rule` - Checks that each target uses a valid PromQL query.
+* `target-rate-interval-rule` - Checks that each target uses $__rate_interval.
+* `target-job-rule` - Checks that every PromQL query has a job matcher.
+* `target-instance-rule` - Checks that every PromQL query has a instance matcher.
 
 ## Exceptions
